Moved desired capabilities from remote options
