+++ page_objects/capabilities.py
@@ -1,5 +1,7 @@
 import os
 from conf import screenshot_conf
+from datetime import datetime
+
 class Capabilities:    
     def saucelab_credentials(self, sauce_options,username,password):
         """Set saucelab credentials."""
@@ -49,7 +51,35 @@ class Capabilities:
         desired_capabilities['app'] = self.browser_stack_upload(app_name, app_path) #upload the application to the Browserstack Storage
         desired_capabilities['bstack:options'] = bstack_mobile_options
         return desired_capabilities
-
+        
     def browserstack_snapshots(self, desired_capabilities):
         desired_capabilities['debug'] = str(screenshot_conf.BS_ENABLE_SCREENSHOTS).lower()
+        return desired_capabilities
+
+    def set_os(desired_capabilities, os_name, os_version):
+        """Set os name and os_version."""      
+        desired_capabilities['os'] = os_name
+        desired_capabilities['osVersion'] = os_version
+
+        return desired_capabilities
+
+    def remote_project_name(desired_capabilities, remote_project_name):
+        """Set remote project name for browserstack."""
+        desired_capabilities['projectName'] = remote_project_name
+
+        return desired_capabilities
+
+    def remote_build_name(desired_capabilities, remote_build_name):
+        """Set remote build name for browserstack."""
+        desired_capabilities['buildName'] = remote_build_name+"_"+str(datetime.now().strftime("%c"))
+
+        return desired_capabilities
+
+    def set_mobile_device(mobile_os_name, mobile_os_version, device_name):
+        """Setup the mobile device."""
+        desired_capabilities = {}
+        desired_capabilities['platformName'] = mobile_os_name
+        desired_capabilities['platformVersion'] = mobile_os_version
+        desired_capabilities['deviceName'] = device_name
+
         return desired_capabilities
